No need to claim we can have such large queues
Also, fix signed vs unsigned comparison warning

+++ src/boundedqueue.h
@@ -40,10 +40,10 @@ template <class T, class T2 = uint64_t>
 class bqueue {
     //Only stores info for N elements
     vector<T>  elems;
-    size_t first;
-    size_t last;
-    size_t maxsize; //max number of history elements
-    size_t queuesize; // Number of current elements (must be < maxsize !)
+    uint32_t first;
+    uint32_t last;
+    uint32_t maxsize; //max number of history elements
+    uint32_t queuesize; // Number of current elements (must be < maxsize !)
     T2  sumofqueue;
     #ifdef STATS_NEEDED
     AvgCalc<T, T2> longTermAvg;
@@ -107,7 +107,7 @@ public:
 
     T prev(int32_t p) const
     {
-        if (p > queuesize)
+        if (p > (int32_t)queuesize)
             return 0;
 
         uint32_t e;
